+++ src/libtcod/sdl2/gl2_display.h
@@ -45,6 +45,7 @@ class OpenGL2Display: public WindowedDisplay {
       int window_flags,
       const std::string& title);
   virtual void set_tileset(std::shared_ptr<Tileset> tileset) override;
+  virtual void accumulate(const TCOD_Console*) override;
   virtual void present(const TCOD_Console*) override;
   virtual auto read_pixels() const -> Image override;
  private: